Rename g3d_file to G3DFile
